requirements.txt
streamlit
numpy
pillow
matplotlib
azure-storage-blob
python-dotenv
pandas
plotly